# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Server API",
        "description": "API сайта",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://siveno.shop"
        }
    ],
    "paths": {
...
        "/api/sizes": {
            "get": {
                "tags": [
                    "Sizes"
                ],
                "summary": "Получение списка размеров",
                "operationId": "207bfad55ed54bbb539320f30845d311",
                "responses": {
                    "200": {
                        "description": "Список размеров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "M"
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "m"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Sizes"
                ],
                "summary": "Создание нового размера",
                "operationId": "82c4cc7818480d67c9cc752e816f2402",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "m"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Размер создан"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/sizes/{slug}/products": {
            "get": {
                "tags": [
                    "Sizes"
                ],
                "summary": "Получение всех товаров по размеру",
                "operationId": "12f803212131e53e5b9a95a0fdc617d6",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug размера",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Футболка"
                                            },
                                            "price": {
                                                "type": "number",
                                                "example": 19.989999999999998436805981327779591083526611328125
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Размер не найден"
                    }
                }
            }
        },
        "/api/sizes/{id}": {
            "delete": {
                "tags": [
                    "Sizes"
                ],
                "summary": "Удаление размера",
                "operationId": "9deca940294a87bef005757011799ccf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID размера",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Размер удален"
                    },
                    "404": {
                        "description": "Размер не найден"
                    }
                }
            }
        },
        "/api/stories": {
            "get": {
                "tags": [
                    "Stories"
                ],
                "summary": "Получить список сторисов с пагинацией",
                "operationId": "85a8aaaa2e289a5e71d86110ca157a93",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список сторисов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Story"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Stories"
                ],
                "summary": "Создать новый сторис",
                "operationId": "f6f11b14595376c9b47f393c5dfc6d0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image_url"
                                ],
                                "properties": {
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/story.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Сторис создан"
                    }
                }
            }
        },
        "/api/stories/{id}": {
            "get": {
                "tags": [
                    "Stories"
                ],
                "summary": "Получить конкретный сторис",
                "operationId": "7551e103bbb8b65b7b1127a79134ed32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о сторисе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Story"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stories"
                ],
                "summary": "Удалить сторис",
                "operationId": "58cdaab9fbacdbe3ec77dc5c5563500e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сторис удален"
                    }
                }
            }
        },
        "/api/stories/{story}/products": {
            "post": {
                "tags": [
                    "Stories"
                ],
                "summary": "Добавить товар в сторис",
                "operationId": "c752a503762f6949ed035820ec5fa140",
                "parameters": [
                    {
                        "name": "story",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар добавлен"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stories"
                ],
                "summary": "Удалить товар из сториса",
                "operationId": "9824a2d1dad300d0fffedf8c979d6745",
                "parameters": [
                    {
                        "name": "story",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар удален"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Модель продукта",
                "required": [
                    "id_product",
                    "id_product_1c",
                    "title",
                    "article",
                    "description",
                    "unit",
                    "image_urls",
                    "price",
                    "weight",
                    "length",
                    "width",
                    "height",
                    "ready"
                ],
                "properties": {
                    "id_product": {
                        "description": "Уникальный идентификатор товара на сайте",
                        "type": "integer",
                        "example": 1
                    },
                    "id_product_1c": {
                        "description": "Уникальный идентификатор товара в 1С",
                        "type": "string",
                        "example": "1c-001"
                    },
                    "title": {
                        "description": "Наименование товара",
                        "type": "string",
                        "example": "Футболка"
                    },
                    "article": {
                        "description": "Артикул товара",
                        "type": "string",
                        "example": "ABC123"
                    },
                    "description": {
                        "description": "Описание товара",
                        "type": "string",
                        "example": "Полное описание товара"
                    },
                    "unit": {
                        "description": "Единица измерения",
                        "type": "string",
                        "example": "шт"
                    },
                    "image_urls": {
                        "description": "Ссылки на изображения",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "https://example.com/image.jpg"
                        }
                    },
                    "price": {
                        "description": "Цена продажи",
                        "type": "number",
                        "format": "float",
                        "example": 199.990000000000009094947017729282379150390625
                    },
                    "weight": {
                        "description": "Вес в килограммах",
                        "type": "number",
                        "format": "float",
                        "example": 0.5
                    },
                    "length": {
                        "description": "Длина в сантиметрах",
                        "type": "number",
                        "format": "float",
                        "example": 30
                    },
                    "width": {
                        "description": "Ширина в сантиметрах",
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "height": {
                        "description": "Высота в сантиметрах",
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "ready": {
                        "description": "Флаг выгрузки/изменения карты товара в 1С",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "Story": {
                "title": "Story",
                "description": "Модель сториса",
                "properties": {
                    "id": {
                        "description": "ID сториса",
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "description": "Ссылка на изображение",
                        "type": "string",
                        "example": "https://example.com/story.jpg"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате Bearer {your_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}